+++ druid/src/scroll_component.rs
@@ -17,11 +17,10 @@
 
 use std::time::Duration;
 
-use crate::kurbo::{Affine, Point, Rect, RoundedRect, Size, Vec2};
+use crate::kurbo::{Point, Rect, Vec2};
 use crate::theme;
-use crate::{
-    Env, Event, EventCtx, LifeCycle, LifeCycleCtx, PaintCtx, Region, RenderContext, TimerToken,
-};
+use crate::widget::Viewport;
+use crate::{Env, Event, EventCtx, LifeCycle, LifeCycleCtx, PaintCtx, RenderContext, TimerToken};
 
 //TODO: Add this to env
 /// Minimum length for any scrollbar to be when measured on that
@@ -63,9 +62,37 @@ pub enum BarHeldState {
     Horizontal(f64),
 }
 
-/// Backing struct for storing scrollbar state
+/// Embeddable component exposing reusable scroll handling logic.
+///
+/// In most situations composing [`Scroll`] is a better idea
+/// for general UI construction. However some cases are not covered by
+/// composing those widgets, such as when a widget needs fine grained
+/// control over its scrolling state or doesn't make sense to exist alone
+/// without scrolling behavior.
+///
+/// `ScrollComponent` contains the input-handling and scrollbar-positioning logic used by
+/// [`Scroll`].  It can be used to add this logic to a custom widget when the need arises.
+///
+/// It can be used like this:
+/// - Store an instance of `ScrollComponent` in your widget's struct, and wrap the child widget to
+///   be scrolled in a [`ClipBox`].
+/// - Call [`event`] and [`lifecycle`] with all event and lifecycle events before propagating them
+///   to children.
+/// - Call [`handle_scroll`] with all events after handling / propagating them.
+/// - Call [`draw_bars`] to draw the scrollbars.
+///
+/// Taking a look at the [`Scroll`] source code can be helpful. You can also do scrolling
+/// without wrapping a child in a [`ClipBox`], but you will need to do certain event and
+/// paint transformations yourself; see the [`ClipBox`] source code for an example.
+///
+/// [`Scroll`]: ../widget/struct.Scroll.html
+/// [`List`]: ../widget/struct.List.html
+/// [`ClipBox`]: ../widget/struct.ClipBox.html
+/// [`event`]: struct.ScrollComponent.html#method.event
+/// [`handle_scroll`]: struct.ScrollComponent.html#method.handle_scroll
+/// [`lifecycle`]: struct.ScrollComponent.html#method.lifecycle
 #[derive(Debug, Copy, Clone)]
-pub struct ScrollbarsState {
+pub struct ScrollComponent {
     /// Current opacity for both scrollbars
     pub opacity: f64,
     /// ID for the timer which schedules scrollbar fade out
@@ -76,7 +103,7 @@ pub struct ScrollbarsState {
     pub held: BarHeldState,
 }
 
-impl Default for ScrollbarsState {
+impl Default for ScrollComponent {
     fn default() -> Self {
         Self {
             opacity: 0.0,
@@ -87,87 +114,15 @@ impl Default for ScrollbarsState {
     }
 }
 
-impl ScrollbarsState {
-    /// true if either scrollbar is currently held down/being dragged
-    pub fn are_held(&self) -> bool {
-        !matches!(self.held, BarHeldState::None)
-    }
-}
-
-/// Embeddable component exposing reusable scroll handling logic.
-///
-/// In most situations composing [`Scroll`] or [`List`] is a better idea
-/// for general UI construction. However some cases are not covered by
-/// composing those widgets, such as when a widget needs fine grained
-/// control over its scrolling state or doesn't make sense to exist alone
-/// without scrolling behavior.
-///
-/// `ScrollComponent` contains the unified and consistent scroll logic
-/// used by both [`Scroll`] and [`List`]. This can be used to add this
-/// logic to a custom widget when the need arises.
-///
-/// It should be used like this:
-/// - Store an instance of `ScrollComponent` in your widget's struct.
-/// - During layout, set the [`content_size`] field to the child's size.
-/// - Call [`event`] and [`lifecycle`] with all event and lifecycle events before propagating them to children.
-/// - Call [`handle_scroll`] with all events after handling / propagating them.
-/// - And finally perform painting using the provided [`paint_content`] function.
-///
-/// Also, taking a look at the [`Scroll`] source code can be helpful.
-///
-/// [`Scroll`]: ../widget/struct.Scroll.html
-/// [`List`]: ../widget/struct.List.html
-/// [`content_size`]: struct.ScrollComponent.html#structfield.content_size
-/// [`event`]: struct.ScrollComponent.html#method.event
-/// [`handle_scroll`]: struct.ScrollComponent.html#method.handle_scroll
-/// [`lifecycle`]: struct.ScrollComponent.html#method.lifecycle
-/// [`paint_content`]: struct.ScrollComponent.html#method.paint_content
-#[derive(Debug, Copy, Clone)]
-pub struct ScrollComponent {
-    /// The size of the scrollable content, make sure to keep up this
-    /// accurate to the content being scrolled
-    pub content_size: Size,
-    /// Current offset of the scrolling content
-    pub scroll_offset: Vec2,
-    /// Current state of both scrollbars
-    pub scrollbars: ScrollbarsState,
-}
-
-impl Default for ScrollComponent {
-    fn default() -> Self {
-        ScrollComponent::new()
-    }
-}
-
 impl ScrollComponent {
     /// Constructs a new [`ScrollComponent`](struct.ScrollComponent.html) for use.
     pub fn new() -> ScrollComponent {
-        ScrollComponent {
-            content_size: Size::default(),
-            scroll_offset: Vec2::new(0.0, 0.0),
-            scrollbars: ScrollbarsState::default(),
-        }
+        Default::default()
     }
 
-    /// Scroll `delta` units.
-    ///
-    /// Returns `true` if the scroll offset has changed.
-    pub fn scroll(&mut self, delta: Vec2, layout_size: Size) -> bool {
-        let mut offset = self.scroll_offset + delta;
-        offset.x = offset
-            .x
-            .min(self.content_size.width - layout_size.width)
-            .max(0.0);
-        offset.y = offset
-            .y
-            .min(self.content_size.height - layout_size.height)
-            .max(0.0);
-        if (offset - self.scroll_offset).hypot2() > 1e-12 {
-            self.scroll_offset = offset;
-            true
-        } else {
-            false
-        }
+    /// true if either scrollbar is currently held down/being dragged
+    pub fn are_bars_held(&self) -> bool {
+        !matches!(self.held, BarHeldState::None)
     }
 
     /// Makes the scrollbars visible, and resets the fade timer.
@@ -175,111 +130,114 @@ impl ScrollComponent {
     where
         F: FnOnce(Duration) -> TimerToken,
     {
-        self.scrollbars.opacity = env.get(theme::SCROLLBAR_MAX_OPACITY);
+        self.opacity = env.get(theme::SCROLLBAR_MAX_OPACITY);
         let fade_delay = env.get(theme::SCROLLBAR_FADE_DELAY);
         let deadline = Duration::from_millis(fade_delay);
-        self.scrollbars.timer_id = request_timer(deadline);
+        self.timer_id = request_timer(deadline);
     }
 
-    /// Calculates the paint rect of the vertical scrollbar.
-    ///
-    /// Returns `Rect::ZERO` if the vertical scrollbar is not visible.
-    pub fn calc_vertical_bar_bounds(&self, viewport: Rect, env: &Env) -> Rect {
-        if viewport.height() >= self.content_size.height {
-            return Rect::ZERO;
+    /// Calculates the paint rect of the vertical scrollbar, or `None` if the vertical scrollbar is
+    /// not visible.
+    pub fn calc_vertical_bar_bounds(&self, port: &Viewport, env: &Env) -> Option<Rect> {
+        let viewport_size = port.rect.size();
+        let content_size = port.content_size;
+        let scroll_offset = port.rect.origin().to_vec2();
+
+        if viewport_size.height >= content_size.height {
+            return None;
         }
 
         let bar_width = env.get(theme::SCROLLBAR_WIDTH);
         let bar_pad = env.get(theme::SCROLLBAR_PAD);
 
-        let percent_visible = viewport.height() / self.content_size.height;
-        let percent_scrolled =
-            self.scroll_offset.y / (self.content_size.height - viewport.height());
+        let percent_visible = viewport_size.height / content_size.height;
+        let percent_scrolled = scroll_offset.y / (content_size.height - viewport_size.height);
 
-        let length = (percent_visible * viewport.height()).ceil();
+        let length = (percent_visible * viewport_size.height).ceil();
         let length = length.max(SCROLLBAR_MIN_SIZE);
 
         let vertical_padding = bar_pad + bar_pad + bar_width;
 
         let top_y_offset =
-            ((viewport.height() - length - vertical_padding) * percent_scrolled).ceil();
+            ((viewport_size.height - length - vertical_padding) * percent_scrolled).ceil();
         let bottom_y_offset = top_y_offset + length;
 
-        let x0 = self.scroll_offset.x + viewport.width() - bar_width - bar_pad;
-        let y0 = self.scroll_offset.y + top_y_offset + bar_pad;
+        let x0 = scroll_offset.x + viewport_size.width - bar_width - bar_pad;
+        let y0 = scroll_offset.y + top_y_offset + bar_pad;
 
-        let x1 = self.scroll_offset.x + viewport.width() - bar_pad;
-        let y1 = self.scroll_offset.y + bottom_y_offset;
+        let x1 = scroll_offset.x + viewport_size.width - bar_pad;
+        let y1 = scroll_offset.y + bottom_y_offset;
 
-        Rect::new(x0, y0, x1, y1)
+        Some(Rect::new(x0, y0, x1, y1))
     }
 
-    /// Calculates the paint rect of the horizontal scrollbar.
-    ///
-    /// Returns `Rect::ZERO` if the horizontal scrollbar is not visible.
-    pub fn calc_horizontal_bar_bounds(&self, viewport: Rect, env: &Env) -> Rect {
-        if viewport.width() >= self.content_size.width {
-            return Rect::ZERO;
+    /// Calculates the paint rect of the horizontal scrollbar, or `None` if the horizontal
+    /// scrollbar is not visible.
+    pub fn calc_horizontal_bar_bounds(&self, port: &Viewport, env: &Env) -> Option<Rect> {
+        let viewport_size = port.rect.size();
+        let content_size = port.content_size;
+        let scroll_offset = port.rect.origin().to_vec2();
+
+        if viewport_size.width >= content_size.width {
+            return None;
         }
 
         let bar_width = env.get(theme::SCROLLBAR_WIDTH);
         let bar_pad = env.get(theme::SCROLLBAR_PAD);
 
-        let percent_visible = viewport.width() / self.content_size.width;
-        let percent_scrolled = self.scroll_offset.x / (self.content_size.width - viewport.width());
+        let percent_visible = viewport_size.width / content_size.width;
+        let percent_scrolled = scroll_offset.x / (content_size.width - viewport_size.width);
 
-        let length = (percent_visible * viewport.width()).ceil();
+        let length = (percent_visible * viewport_size.width).ceil();
         let length = length.max(SCROLLBAR_MIN_SIZE);
 
         let horizontal_padding = bar_pad + bar_pad + bar_width;
 
         let left_x_offset =
-            ((viewport.width() - length - horizontal_padding) * percent_scrolled).ceil();
+            ((viewport_size.width - length - horizontal_padding) * percent_scrolled).ceil();
         let right_x_offset = left_x_offset + length;
 
-        let x0 = self.scroll_offset.x + left_x_offset + bar_pad;
-        let y0 = self.scroll_offset.y + viewport.height() - bar_width - bar_pad;
+        let x0 = scroll_offset.x + left_x_offset + bar_pad;
+        let y0 = scroll_offset.y + viewport_size.height - bar_width - bar_pad;
 
-        let x1 = self.scroll_offset.x + right_x_offset;
-        let y1 = self.scroll_offset.y + viewport.height() - bar_pad;
+        let x1 = scroll_offset.x + right_x_offset;
+        let y1 = scroll_offset.y + viewport_size.height - bar_pad;
 
-        Rect::new(x0, y0, x1, y1)
+        Some(Rect::new(x0, y0, x1, y1))
     }
 
     /// Draw scroll bars.
-    pub fn draw_bars(&self, ctx: &mut PaintCtx, viewport: Rect, env: &Env) {
-        if self.scrollbars.opacity <= 0.0 {
+    pub fn draw_bars(&self, ctx: &mut PaintCtx, port: &Viewport, env: &Env) {
+        let scroll_offset = port.rect.origin().to_vec2();
+        if self.opacity <= 0.0 {
             return;
         }
 
-        let brush = ctx.render_ctx.solid_brush(
-            env.get(theme::SCROLLBAR_COLOR)
-                .with_alpha(self.scrollbars.opacity),
-        );
+        let brush = ctx
+            .render_ctx
+            .solid_brush(env.get(theme::SCROLLBAR_COLOR).with_alpha(self.opacity));
         let border_brush = ctx.render_ctx.solid_brush(
             env.get(theme::SCROLLBAR_BORDER_COLOR)
-                .with_alpha(self.scrollbars.opacity),
+                .with_alpha(self.opacity),
         );
 
         let radius = env.get(theme::SCROLLBAR_RADIUS);
         let edge_width = env.get(theme::SCROLLBAR_EDGE_WIDTH);
 
         // Vertical bar
-        if viewport.height() < self.content_size.height {
-            let bounds = self
-                .calc_vertical_bar_bounds(viewport, env)
-                .inset(-edge_width / 2.0);
-            let rect = RoundedRect::from_rect(bounds, radius);
+        if let Some(bounds) = self.calc_vertical_bar_bounds(port, env) {
+            let rect = (bounds - scroll_offset)
+                .inset(-edge_width / 2.0)
+                .to_rounded_rect(radius);
             ctx.render_ctx.fill(rect, &brush);
             ctx.render_ctx.stroke(rect, &border_brush, edge_width);
         }
 
         // Horizontal bar
-        if viewport.width() < self.content_size.width {
-            let bounds = self
-                .calc_horizontal_bar_bounds(viewport, env)
-                .inset(-edge_width / 2.0);
-            let rect = RoundedRect::from_rect(bounds, radius);
+        if let Some(bounds) = self.calc_horizontal_bar_bounds(port, env) {
+            let rect = (bounds - scroll_offset)
+                .inset(-edge_width / 2.0)
+                .to_rounded_rect(radius);
             ctx.render_ctx.fill(rect, &brush);
             ctx.render_ctx.stroke(rect, &border_brush, edge_width);
         }
@@ -288,11 +246,13 @@ impl ScrollComponent {
     /// Tests if the specified point overlaps the vertical scrollbar
     ///
     /// Returns false if the vertical scrollbar is not visible
-    pub fn point_hits_vertical_bar(&self, viewport: Rect, pos: Point, env: &Env) -> bool {
-        if viewport.height() < self.content_size.height {
+    pub fn point_hits_vertical_bar(&self, port: &Viewport, pos: Point, env: &Env) -> bool {
+        let viewport_size = port.rect.size();
+        let scroll_offset = port.rect.origin().to_vec2();
+
+        if let Some(mut bounds) = self.calc_vertical_bar_bounds(port, env) {
             // Stretch hitbox to edge of widget
-            let mut bounds = self.calc_vertical_bar_bounds(viewport, env);
-            bounds.x1 = self.scroll_offset.x + viewport.width();
+            bounds.x1 = scroll_offset.x + viewport_size.width;
             bounds.contains(pos)
         } else {
             false
@@ -302,11 +262,13 @@ impl ScrollComponent {
     /// Tests if the specified point overlaps the horizontal scrollbar
     ///
     /// Returns false if the horizontal scrollbar is not visible
-    pub fn point_hits_horizontal_bar(&self, viewport: Rect, pos: Point, env: &Env) -> bool {
-        if viewport.width() < self.content_size.width {
+    pub fn point_hits_horizontal_bar(&self, port: &Viewport, pos: Point, env: &Env) -> bool {
+        let viewport_size = port.rect.size();
+        let scroll_offset = port.rect.origin().to_vec2();
+
+        if let Some(mut bounds) = self.calc_horizontal_bar_bounds(port, env) {
             // Stretch hitbox to edge of widget
-            let mut bounds = self.calc_horizontal_bar_bounds(viewport, env);
-            bounds.y1 = self.scroll_offset.y + viewport.height();
+            bounds.y1 = scroll_offset.y + viewport_size.height;
             bounds.contains(pos)
         } else {
             false
@@ -316,38 +278,43 @@ impl ScrollComponent {
     /// Checks if the event applies to the scroll behavior, uses it, and marks it handled
     ///
     /// Make sure to call on every event
-    pub fn event(&mut self, ctx: &mut EventCtx, event: &Event, env: &Env) {
-        let size = ctx.size();
-        let viewport = size.to_rect();
+    pub fn event(&mut self, port: &mut Viewport, ctx: &mut EventCtx, event: &Event, env: &Env) {
+        let viewport_size = port.rect.size();
+        let content_size = port.content_size;
+        let scroll_offset = port.rect.origin().to_vec2();
 
         let scrollbar_is_hovered = match event {
             Event::MouseMove(e) | Event::MouseUp(e) | Event::MouseDown(e) => {
-                let offset_pos = e.pos + self.scroll_offset;
-                self.point_hits_vertical_bar(viewport, offset_pos, env)
-                    || self.point_hits_horizontal_bar(viewport, offset_pos, env)
+                let offset_pos = e.pos + scroll_offset;
+                self.point_hits_vertical_bar(port, offset_pos, env)
+                    || self.point_hits_horizontal_bar(port, offset_pos, env)
             }
             _ => false,
         };
 
-        if self.scrollbars.are_held() {
+        if self.are_bars_held() {
             // if we're dragging a scrollbar
             match event {
                 Event::MouseMove(event) => {
-                    match self.scrollbars.held {
+                    match self.held {
                         BarHeldState::Vertical(offset) => {
-                            let scale_y = viewport.height() / self.content_size.height;
-                            let bounds = self.calc_vertical_bar_bounds(viewport, env);
-                            let mouse_y = event.pos.y + self.scroll_offset.y;
+                            let scale_y = viewport_size.height / content_size.height;
+                            let bounds = self
+                                .calc_vertical_bar_bounds(port, env)
+                                .unwrap_or(Rect::ZERO);
+                            let mouse_y = event.pos.y + scroll_offset.y;
                             let delta = mouse_y - bounds.y0 - offset;
-                            self.scroll(Vec2::new(0f64, (delta / scale_y).ceil()), size);
+                            port.pan_by(Vec2::new(0f64, (delta / scale_y).ceil()));
                             ctx.set_handled();
                         }
                         BarHeldState::Horizontal(offset) => {
-                            let scale_x = viewport.width() / self.content_size.width;
-                            let bounds = self.calc_horizontal_bar_bounds(viewport, env);
-                            let mouse_x = event.pos.x + self.scroll_offset.x;
+                            let scale_x = viewport_size.height / content_size.width;
+                            let bounds = self
+                                .calc_horizontal_bar_bounds(port, env)
+                                .unwrap_or(Rect::ZERO);
+                            let mouse_x = event.pos.x + scroll_offset.x;
                             let delta = mouse_x - bounds.x0 - offset;
-                            self.scroll(Vec2::new((delta / scale_x).ceil(), 0f64), size);
+                            port.pan_by(Vec2::new((delta / scale_x).ceil(), 0f64));
                             ctx.set_handled();
                         }
                         _ => (),
@@ -355,11 +322,11 @@ impl ScrollComponent {
                     ctx.request_paint();
                 }
                 Event::MouseUp(_) => {
-                    self.scrollbars.held = BarHeldState::None;
+                    self.held = BarHeldState::None;
                     ctx.set_active(false);
 
                     if !scrollbar_is_hovered {
-                        self.scrollbars.hovered = BarHoveredState::None;
+                        self.hovered = BarHoveredState::None;
                         self.reset_scrollbar_fade(|d| ctx.request_timer(d), env);
                     }
 
@@ -371,32 +338,34 @@ impl ScrollComponent {
             // if we're over a scrollbar but not dragging
             match event {
                 Event::MouseMove(event) => {
-                    let offset_pos = event.pos + self.scroll_offset;
-                    if self.point_hits_vertical_bar(viewport, offset_pos, env) {
-                        self.scrollbars.hovered = BarHoveredState::Vertical;
-                    } else if self.point_hits_horizontal_bar(viewport, offset_pos, env) {
-                        self.scrollbars.hovered = BarHoveredState::Horizontal;
+                    let offset_pos = event.pos + scroll_offset;
+                    if self.point_hits_vertical_bar(port, offset_pos, env) {
+                        self.hovered = BarHoveredState::Vertical;
+                    } else if self.point_hits_horizontal_bar(port, offset_pos, env) {
+                        self.hovered = BarHoveredState::Horizontal;
                     } else {
                         unreachable!();
                     }
 
-                    self.scrollbars.opacity = env.get(theme::SCROLLBAR_MAX_OPACITY);
-                    self.scrollbars.timer_id = TimerToken::INVALID; // Cancel any fade out in progress
+                    self.opacity = env.get(theme::SCROLLBAR_MAX_OPACITY);
+                    self.timer_id = TimerToken::INVALID; // Cancel any fade out in progress
                     ctx.request_paint();
                     ctx.set_handled();
                 }
                 Event::MouseDown(event) => {
-                    let pos = event.pos + self.scroll_offset;
+                    let pos = event.pos + scroll_offset;
 
-                    if self.point_hits_vertical_bar(viewport, pos, env) {
+                    if self.point_hits_vertical_bar(port, pos, env) {
                         ctx.set_active(true);
-                        self.scrollbars.held = BarHeldState::Vertical(
-                            pos.y - self.calc_vertical_bar_bounds(viewport, env).y0,
+                        self.held = BarHeldState::Vertical(
+                            // The bounds must be non-empty, because the point hits the scrollbar.
+                            pos.y - self.calc_vertical_bar_bounds(port, env).unwrap().y0,
                         );
-                    } else if self.point_hits_horizontal_bar(viewport, pos, env) {
+                    } else if self.point_hits_horizontal_bar(port, pos, env) {
                         ctx.set_active(true);
-                        self.scrollbars.held = BarHeldState::Horizontal(
-                            pos.x - self.calc_horizontal_bar_bounds(viewport, env).x0,
+                        self.held = BarHeldState::Horizontal(
+                            // The bounds must be non-empty, because the point hits the scrollbar.
+                            pos.x - self.calc_horizontal_bar_bounds(port, env).unwrap().x0,
                         );
                     } else {
                         unreachable!();
@@ -412,38 +381,33 @@ impl ScrollComponent {
             match event {
                 Event::MouseMove(_) => {
                     // if we have just stopped hovering
-                    if self.scrollbars.hovered.is_hovered() && !scrollbar_is_hovered {
-                        self.scrollbars.hovered = BarHoveredState::None;
+                    if self.hovered.is_hovered() && !scrollbar_is_hovered {
+                        self.hovered = BarHoveredState::None;
                         self.reset_scrollbar_fade(|d| ctx.request_timer(d), env);
                     }
                 }
-                Event::Timer(id) if *id == self.scrollbars.timer_id => {
+                Event::Timer(id) if *id == self.timer_id => {
                     // Schedule scroll bars animation
                     ctx.request_anim_frame();
-                    self.scrollbars.timer_id = TimerToken::INVALID;
+                    self.timer_id = TimerToken::INVALID;
                     ctx.set_handled();
                 }
                 Event::AnimFrame(interval) => {
                     // Guard by the timer id being invalid, otherwise the scroll bars would fade
                     // immediately if some other widget started animating.
-                    if self.scrollbars.timer_id == TimerToken::INVALID {
+                    if self.timer_id == TimerToken::INVALID {
                         // Animate scroll bars opacity
                         let diff = 2.0 * (*interval as f64) * 1e-9;
-                        self.scrollbars.opacity -= diff;
-                        if self.scrollbars.opacity > 0.0 {
+                        self.opacity -= diff;
+                        if self.opacity > 0.0 {
                             ctx.request_anim_frame();
                         }
 
-                        let viewport = ctx.size().to_rect();
-                        if viewport.width() < self.content_size.width {
-                            ctx.request_paint_rect(
-                                self.calc_horizontal_bar_bounds(viewport, env) - self.scroll_offset,
-                            );
+                        if let Some(bounds) = self.calc_horizontal_bar_bounds(port, env) {
+                            ctx.request_paint_rect(bounds - scroll_offset);
                         }
-                        if viewport.height() < self.content_size.height {
-                            ctx.request_paint_rect(
-                                self.calc_vertical_bar_bounds(viewport, env) - self.scroll_offset,
-                            );
+                        if let Some(bounds) = self.calc_vertical_bar_bounds(port, env) {
+                            ctx.request_paint_rect(bounds - scroll_offset);
                         }
                     }
                 }
@@ -454,10 +418,16 @@ impl ScrollComponent {
     }
 
     /// Applies mousewheel scrolling if the event has not already been handled
-    pub fn handle_scroll(&mut self, ctx: &mut EventCtx, event: &Event, env: &Env) {
+    pub fn handle_scroll(
+        &mut self,
+        port: &mut Viewport,
+        ctx: &mut EventCtx,
+        event: &Event,
+        env: &Env,
+    ) {
         if !ctx.is_handled() {
             if let Event::Wheel(mouse) = event {
-                if self.scroll(mouse.wheel_delta, ctx.size()) {
+                if port.pan_by(mouse.wheel_delta) {
                     ctx.request_paint();
                     ctx.set_handled();
                     self.reset_scrollbar_fade(|d| ctx.request_timer(d), env);
@@ -474,25 +444,5 @@ impl ScrollComponent {
             // Show the scrollbars any time our size changes
             self.reset_scrollbar_fade(|d| ctx.request_timer(d), &env);
         }
-    }
-
-    /// Helper function to paint a closure at the correct offset with clipping and scrollbars
-    pub fn paint_content(
-        self,
-        ctx: &mut PaintCtx,
-        env: &Env,
-        f: impl FnOnce(Region, &mut PaintCtx),
-    ) {
-        let viewport = ctx.size().to_rect();
-        ctx.with_save(|ctx| {
-            ctx.clip(viewport);
-            ctx.transform(Affine::translate(-self.scroll_offset));
-
-            let mut visible = ctx.region().clone();
-            visible += self.scroll_offset;
-            f(visible, ctx);
-
-            self.draw_bars(ctx, viewport, env);
-        });
     }
 }